Greedy meshing initial implementation
greedy meshing WIP

rendering mesh wip

greedy meshing WIP

rework & refactor, to be fixed

fixes and rename constants

greedy meshing wip

compiler errors fix

greedy meshing wip

greedy meshing progress

compiler error fixes, moved test cases to notes

greedy meshing more progress

small compilation fix

saving vertices and indices greedy meshing

std:: added

iota usage added

+++ VoxTerGenEngine/include/world/Block.hpp
@@ -17,8 +17,8 @@ class Block
 {
 private:
     BlockType type_;
-    std::uint32_t sun_light_;
-    std::uint32_t block_light_;
+    std::uint8_t sun_light_;
+    std::uint8_t block_light_;
 
 public:
 	Block(BlockType type = BlockType::Air);
